[Anubhav]Add. Introducing DTO and MOdel to Employee Payroll App

+++ EmployeePayrollApp/src/main/java/com/bridgelabz/EmployeePayrollApp/service/EmployeeService.java
@@ -1,11 +1,13 @@
 package com.bridgelabz.EmployeePayrollApp.service;
 
+import com.bridgelabz.EmployeePayrollApp.dto.EmployeeDTO;
 import com.bridgelabz.EmployeePayrollApp.model.Employee;
 import com.bridgelabz.EmployeePayrollApp.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class EmployeeService {
@@ -14,13 +16,11 @@ public class EmployeeService {
     private EmployeeRepository employeeRepository;
 
     // Create Employee
-    public Employee createEmployee(Employee employee) {
+    public Employee createEmployee(EmployeeDTO employeeDTO) {
+        Employee employee = new Employee();
+        employee.setName(employeeDTO.getName());
+        employee.setSalary(employeeDTO.getSalary());
         return employeeRepository.save(employee);
-    }
-
-    // Get Employee by ID
-    public Employee getEmployeeById(int id) {
-        return employeeRepository.findById(id).orElse(null);
     }
 
     // Get All Employees
@@ -28,19 +28,28 @@ public class EmployeeService {
         return employeeRepository.findAll();
     }
 
+    // Get Employee by ID
+    public Optional<Employee> getEmployeeById(int id) {
+        return employeeRepository.findById(id);
+    }
+
     // Update Employee
-    public Employee updateEmployee(int id, Employee updatedEmployee) {
-        Employee existingEmployee = employeeRepository.findById(id).orElse(null);
-        if (existingEmployee != null) {
-            existingEmployee.setName(updatedEmployee.getName());
-            existingEmployee.setSalary(updatedEmployee.getSalary());
-            return employeeRepository.save(existingEmployee);
-        }
-        return null;
+    public Optional<Employee> updateEmployee(int id, EmployeeDTO employeeDTO) {
+        Optional<Employee> existingEmployee = employeeRepository.findById(id);
+        existingEmployee.ifPresent(employee -> {
+            employee.setName(employeeDTO.getName());
+            employee.setSalary(employeeDTO.getSalary());
+            employeeRepository.save(employee);
+        });
+        return existingEmployee;
     }
 
     // Delete Employee
-    public void deleteEmployee(int id) {
-        employeeRepository.deleteById(id);
+    public boolean deleteEmployee(int id) {
+        if (employeeRepository.existsById(id)) {
+            employeeRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
